Add special tokens for keys in prompt (#39)
This configures the tokenizer so that strings `### Instruction:`, `### Response:`, and `### End` are all represented by a single token ID each.  This simplifies the logic for finding the response.  `generate` is also now configured to stop generation at `### End`, making generation faster.

The default `max_new_tokens` has been double to 256.

The notebook now has widgets `local_training_root` and `dbfs_output_root` for configuring where data is stored locally and in DBFS.  By default if `local_training_root` is not provided it now uses `/local_disk0` if it exists and otherwise defaults to a subdir of the home directory, as before.